refactored error header
